Allow queue persistence while navigate
